feat: configuracion basica del router
