+++ src/pages/home/Home.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Home = () => {
+  return (
+    <div>Home</div>
+  )
+}
+
+export default Home
